spotifAI/data/requirements.txt
Flask==2.0.1
pandas==1.1.5
waitress==2.0.0
spotipy==2.19.0
google-cloud-secret-manager==2.1.0